Início do menu por perfil, close #46
