+++ frontend/pages/students/participation/index.js
@@ -1,0 +1,3 @@
+export default function Participation() {
+  return <>Participation</>
+}
